Airflow_DIP/requirements.txt
tensorflow==1.13.1
apache-airflow==1.10.2